Some cleanup
This is trying to extend and standardize a few things about the
range of notebooks:

- Using a standard color scale for speed map plots, to visually
compare speeds on the circuits

- Creating a more consistent range of plots for the Miami GP,
which is being used as the standard-bearer for analysis. As a
result, we create summary and team/driver pace notebooks for
this race (so that all five notebooks exist for this race).

- There is a sixth notebook that isn't being replicated for
Miami, the Lap driver notebook (currently only part of the
Imola GP), which I'm not certain will be part of the standard
set of notebooks.

- Both races with a summary notebook get a new gear distribution
plot, showing how many measurements are received with the car
in each gear.